+++ tests/test_class_docstrings.py
@@ -103,13 +103,11 @@ def test_abstract_model(app, do_autodoc):
         "",
         "   Relationship fields:",
         "",
-        "   :param simple_model: Simple model "
-        "(related name: :attr:`~dummy_django_app.models.SimpleModel.abstractmodel`)",
+        "   :param simple_model: Simple model",
         "   :type simple_model: :class:`~django.db.models.ForeignKey` to :class:`~SimpleModel`",
-        "   :param user: User (related name: :attr:`~django.contrib.auth.models.User.+`)",
+        "   :param user: User",
         "   :type user: :class:`~django.db.models.ForeignKey` to :class:`~auth.User`",
-        "   :param foreignkey_self: Foreignkey self "
-        "(related name: :attr:`~dummy_django_app.models.AbstractModel.abstractmodel`)",
+        "   :param foreignkey_self: Foreignkey self",
         "   :type foreignkey_self: :class:`~django.db.models.ForeignKey` to :class:`~self`",
         "",
     ]